IterableCondition -> IterableFunction
