+++ core/src/main/java/io/doov/core/dsl/impl/IterableFunction.java
@@ -15,15 +15,25 @@
  */
 package io.doov.core.dsl.impl;
 
+import static io.doov.core.dsl.meta.function.IterableFunctionMetadata.containsMetadata;
+import static io.doov.core.dsl.meta.function.IterableFunctionMetadata.hasNotSizeMetadata;
+import static io.doov.core.dsl.meta.function.IterableFunctionMetadata.hasSizeMetadata;
+import static io.doov.core.dsl.meta.predicate.UnaryPredicateMetadata.isEmptyMetadata;
+import static io.doov.core.dsl.meta.predicate.UnaryPredicateMetadata.isNotEmptyMetadata;
+import static java.util.stream.Collectors.toSet;
+import static java.util.stream.StreamSupport.stream;
+
+import java.util.Arrays;
 import java.util.Optional;
 import java.util.function.BiFunction;
 
 import io.doov.core.FieldModel;
 import io.doov.core.dsl.DslField;
 import io.doov.core.dsl.lang.Context;
+import io.doov.core.dsl.lang.StepCondition;
 import io.doov.core.dsl.meta.predicate.PredicateMetadata;
 
-public class IterableFunction<T, C extends Iterable<T>> extends IterableCondition<T, C> {
+public class IterableFunction<T, C extends Iterable<T>> extends DefaultCondition<C> {
 
     public IterableFunction(DslField<C> field) {
         super(field);
@@ -31,6 +41,37 @@ public class IterableFunction<T, C extends Iterable<T>> extends IterableConditio
 
     public IterableFunction(PredicateMetadata metadata, BiFunction<FieldModel, Context, Optional<C>> value) {
         super(metadata, value);
+    }
+
+    public StepCondition contains(T value) {
+        return LeafStepCondition.stepCondition(containsMetadata(metadata, value), getFunction(),
+                collection -> stream(collection.spliterator(), false).anyMatch(value::equals));
+    }
+
+    @SafeVarargs
+    public final StepCondition containsAll(T... values) {
+        return LeafStepCondition.stepCondition(containsMetadata(metadata, Arrays.asList(values)), getFunction(),
+                iterable -> stream(iterable.spliterator(), false).collect(toSet()).containsAll(Arrays.asList(values)));
+    }
+
+    public StepCondition isEmpty() {
+        return LeafStepCondition.stepCondition(isEmptyMetadata(metadata), getFunction(),
+                iterable -> !iterable.iterator().hasNext());
+    }
+
+    public StepCondition isNotEmpty() {
+        return LeafStepCondition.stepCondition(isNotEmptyMetadata(metadata), getFunction(),
+                iterable -> iterable.iterator().hasNext());
+    }
+
+    public StepCondition hasSize(int size) {
+        return LeafStepCondition.stepCondition(hasSizeMetadata(metadata, size), getFunction(),
+                iterable -> stream(iterable.spliterator(), false).count() == size);
+    }
+
+    public StepCondition hasNotSize(int size) {
+        return LeafStepCondition.stepCondition(hasNotSizeMetadata(metadata, size), getFunction(),
+                iterable -> stream(iterable.spliterator(), false).count() != size);
     }
 
 }
